scripts/sagemaker_model/code/requirements.txt
transformers==4.37.0
Pillow>=9.0.0
torch>=1.13.0
